added argument checking for supabase functions

+++ supabaseTables/instruments.js
@@ -1,4 +1,6 @@
 async function getInstrument (instrumentId, supabaseClient) {
+  if (!instrumentId) throw new Error('No instrumentId provided')
+  if (!supabaseClient) throw new Error('No supabaseClient provided')
   const { data: Instruments, error } = await supabaseClient
     .from('Instruments')
     .select('id, name, status, active')
@@ -9,6 +11,8 @@ async function getInstrument (instrumentId, supabaseClient) {
   return Instruments
 }
 async function getInstruments (songID, supabaseClient) {
+  if (!songID) throw new Error('No songID provided')
+  if (!supabaseClient) throw new Error('No supabaseClient provided')
   const { data: Instruments, error } = await supabaseClient
     .from('Instruments')
     .select('id, name, status')
@@ -20,6 +24,9 @@ async function getInstruments (songID, supabaseClient) {
 }
 
 async function insertInstruments (instrumentName, songID, supabaseClient) {
+  if (!instrumentName) throw new Error('No instrumentName provided')
+  if (!songID) throw new Error('No songID provided')
+  if (!supabaseClient) throw new Error('No supabaseClient provided')
   const { data, error } = await supabaseClient
     .from('Instruments')
     .insert([
@@ -33,6 +40,9 @@ async function insertInstruments (instrumentName, songID, supabaseClient) {
 }
 
 async function updateInstrument (instrumentId, update, supabaseClient) {
+  if (!instrumentId) throw new Error('No instrumentId provided')
+  if (!update) throw new Error('No update provided')
+  if (!supabaseClient) throw new Error('No supabaseClient provided')
   const { data: Status, error } = await supabaseClient
     .from('Instruments')
     .update({ status: `${update}` })
@@ -46,6 +56,8 @@ async function updateInstrument (instrumentId, update, supabaseClient) {
 }
 
 async function deleteInstrument (instrumentId, supabaseClient) {
+  if (!instrumentId) throw new Error('No instrumentId provided')
+  if (!supabaseClient) throw new Error('No supabaseClient provided')
   const { data, error } = await supabaseClient
     .from('Instruments')
     .delete()
